introduced a number of requested changes
